refactor(contents): extract fileFilter to util & add doc
