+++ src/global/utils/index.ts
@@ -1,3 +1,4 @@
+export { fileFilter } from './file-filter';
 export { notUndef } from './not-undefined';
 export { intOrUndefined, floatOrUndefined } from './number-or-undefined';
 export { validateNumber } from './validate-number';
